Small fixes to the various random vector functions
